requirements.txt
DjangoAsyncAdmin==6.11.0
djangorestframework==3.14.0
elasticsearch==8.4.3
minio==7.1.11
Pygments==2.11.2
requests==2.28.1
docutils==0.19
concurrent-log-handler==0.9.20
django-log-request-id==2.0.0
pyjwt==2.6.0
django-photologue~=3.15.1
django-sortedm2m~=3.1.1
django-revproxy~=0.12.0
opencv-python
django-crontab
pyicloud==1.0.0
piexif==1.1.3
psycopg2-binary
ffmpeg-python==0.2.0
moviepy==1.0.3